refactor(frontend): fix player and sessions bugs
