+++ rust-srec/frontend/src/server/tokenRefresh.ts
@@ -9,11 +9,8 @@
  */
 
 import type { ClientSessionData, SessionData } from '../utils/session';
-import {
-  sanitizeClientSession,
-  isValidSession,
-  useAppSession,
-} from '../utils/session';
+import { sanitizeClientSession, isValidSession } from '../utils/session';
+import { useAppSession } from '../utils/session.server';
 import { BASE_URL } from '../utils/env';
 
 type RefreshOutcome = {
